requirements.txt
python-dotenv
openai
PyPDF2
langchain
langchain-astradb
astrapy
requests